+++ client/src/definitions/navigation.d.ts
@@ -1,5 +1,3 @@
-import type { string } from "yup";
-
 export interface NavItem {
   label: string;
   href: string;
