+++ client/src/lib/components/Footer/Footer.svelte
