+++ client/src/routes/__layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import Breadcrumbs from "$lib/components/Breadcrumbs.svelte";
+  import Breadcrumbs from "$lib/components/Breadcrumbs/Breadcrumbs.svelte";
   import Header from "$lib/components/Header/Header.svelte";
-  import Footer from "$lib/footer/Footer.svelte";
+  import Footer from "$lib/components/Footer/Footer.svelte";
   import { onMount } from "svelte";
   import { page } from "$app/stores";
   import "../app.css";
